视频对课程时长和观看次数的维护

+++ classroom-pojo/src/main/java/com/dingxin/pojo/po/Curriculum.java
@@ -90,7 +90,7 @@ public class Curriculum extends Model<Curriculum> {
     /**
      * 观看次数(课程下的观看次数为该课程下所有视频的观看次数)
      */
-    private Integer watchAmount;
+    private Long watchAmount;
     /**
      * 删除标志(目前采用假删除的方式)
      */
